Reorganized code structure; Updated README
