+++ README.md
@@ -1,18 +1,35 @@
-# ColorTaiko
+# ColorTaiko!
 
-iml project - web dev
+**ColorTaiko!** is an IML (Illinois Mathematics Lab) project that began in Spring 2024 in the Mathematics Department at the University of Illinois Urbana-Champaign (UIUC) and continues into Fall 2024. It introduces a playful, interactive game that challenges players to explore color patterns and solve puzzles. This preliminary version (currently level 1, suitable for beginners) is just the beginning, and we will continue to expand and improve the game in the coming semesters. Future levels will be added, so come back and try your skills as the game evolves!
 
+## What is ColorTaiko!?
 
-YOOOOOOOOOO
+The game is inspired by the research article **"The topology and geometry of units and zero-divisors: origami"** by Igor Mineyev. This work explores how certain colorings of bipartite graphs could potentially lead to counterexamples to the **Kaplansky conjectures**, long-standing open problems in algebra related to group algebras. In the full version of the game, solving increasingly difficult levels could potentially bring new insights or counterexamples related to these mathematical problems. 
 
-this is change from maxwell
+In other words, **ColorTaiko!** is not just a game but an interactive representation of complex mathematical concepts that may contribute to solving important algebraic conjectures. If you can play and succeed at future, more challenging levels, you might actually help demonstrate a counterexample for these famous problems.
 
-CAN YOU SEE THIS
+## Background and Motivation
 
-Hey its ila!
+The **Kaplansky conjectures** are significant open questions in the field of algebra, and they relate to properties of group algebras. **ColorTaiko!** takes inspiration from these problems and presents them in a fun, engaging format. The idea is that by playing through each level, players explore mathematical structures and engage with an interactive representation of these conjectures.
 
+In Fall 2023, Igor Mineyev submitted a grant application to the National Science Foundation (NSF) Topology program, seeking support for this project, the underlying research, and the people involved. Unfortunately, NSF did not support the proposal. Nonetheless, this preliminary version of **ColorTaiko!** was created through dedicated efforts, and we continue to improve it.
 
-this is a change from bhavika
+## Current Version
 
-HELLO
+The current version of **ColorTaiko!** is a kindergarten-level introduction to the game's concepts. It allows players to explore the initial concepts with a simple, user-friendly interface. No prior explanation is given—players are encouraged to dive in and discover the mechanics on their own.
 
+### How to Play
+
+Without any instructions provided, you’re invited to experiment and discover how to play **ColorTaiko!** The game is designed to be intuitive, so feel free to explore the gameplay and see if you can figure out the mechanics on your own!
+
+### Future Developments
+
+We plan to add more challenging levels in the future, each one incrementally bringing players closer to the core mathematical ideas behind the game. Level 1 is only the beginning, so be sure to check back regularly to try new levels as they are added!
+
+## Acknowledgments
+
+This project is powered by the passion and efforts of the UIUC Mathematics Department's IML, and despite the lack of NSF support, we continue to move forward with enthusiasm and creativity.
+
+## More Information
+
+Additional information, updates, and resources will be added here as the project progresses. Stay tuned for new insights, level releases, and research developments. Thank you for being a part of the journey with **ColorTaiko!**
