+++ client/index.html
@@ -2,11 +2,10 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link rel="icon" type="image/png" href="./src/assets/taiko_picture-for-IML.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>ColorTaiko!</title>
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=STIX+Two+Math&display=swap">
-    <link href="https://fonts.googleapis.com/css2?family=Cinzel+Decorative:wght@700&display=swap" rel="stylesheet">
+    <link rel="stylesheet" href="/src/styles/global.css">
   </head>
   <body>
     <div id="root"></div>
