+++ client/src/App.jsx
@@ -1,22 +1,23 @@
 import { useState, useRef, useEffect } from "react";
-import InputBox from "./InputBox";
-import TaikoNode from "./TaikoNode";
-import ErrorModal from "./ErrorModal";
-import LargeArcEdge from "./LargeArcEdge";
+import TaikoNode from "./components/TaikoNodes/TaikoNode";
+import ErrorModal from "./components/ErrorModal";
 
 import { generateColor } from "./utils/colorUtils";
 import { drawConnections } from "./utils/drawingUtils";
 import { checkAndGroupConnections } from "./utils/MergeUtils";
+import { calculateProgress } from "./utils/calculateProgress";
+import { checkAndAddNewNodes} from "./utils/checkAndAddNewNodes";
 
-import clickSound from "./assets/sound effect/Click.wav";
-import errorSound from "./assets/sound effect/Error.wav";
-import connectSound from "./assets/sound effect/Connection.wav";
 import SettingIconImage from "./assets/setting-icon.png";
-import { SettingsMenu } from "./utils/settingMenu";
 
-const edgeTypes = {
-  custom: LargeArcEdge, // Register custom arc edge type
-};
+import SettingsMenu from "./components/ToolMenu/settingMenu";
+import ProgressBar from "./components/ProgressBar/progressBar";
+import Title from "./components/title";
+
+import { useAudio } from './hooks/useAudio';
+import { useSettings } from './hooks/useSetting';
+
+
 
 function App() {
   const [topRowCount, setTopRowCount] = useState(1);
@@ -24,86 +25,45 @@ function App() {
   const [showNodes, setShowNodes] = useState(true);
   const [selectedNodes, setSelectedNodes] = useState([]);
   const [connections, setConnections] = useState([]);
+  const [connectionPairs, setConnectionPairs] = useState([]);
+  const [connectionGroups, setConnectionGroups] = useState([]);
   const [edgeState, setEdgeState] = useState(null);
+  const [progress, setProgress] = useState(0);
+  const [currentColor, setCurrentColor] = useState(0);
   const [errorMessage, setErrorMessage] = useState("");
   const svgRef = useRef(null);
-
-  const [progress, setProgress] = useState(0);
-
-  const clickAudio = new Audio(clickSound);
-  const errorAudio = new Audio(errorSound);
-  const connectAudio = new Audio(connectSound);
-
-  const [tooltipVisible, setTooltipVisible] = useState(false);
-  const [tooltipPosition, setTooltipPosition] = useState({ x: 0, y: 0 });
-
-  const [showSettings, setShowSettings] = useState(false);
-  const [offset, setOffset] = useState(() => {
-    const savedOffset = localStorage.getItem("offset");
-    return savedOffset !== null ? parseInt(savedOffset, 10) : 5;
-  });
-  const [soundBool, setSoundBool] = useState(() => {
-    const savedSound = localStorage.getItem("sound");
-    return savedSound !== null ? JSON.parse(savedSound) : true;
-  });
-
-  const [blackDotEffect, setBlackDotEffect] = useState(() => {
-    const savedDotEffectState = localStorage.getItem("blackDotEffect");
-    return savedDotEffectState ? JSON.parse(savedDotEffectState) : false;
-  });
-
-  useEffect(() => {
-    localStorage.setItem("sound", soundBool);
-  }, [soundBool]);
-
-  useEffect(() => {
-    localStorage.setItem("blackDotEffect", JSON.stringify(blackDotEffect));
-  }, [blackDotEffect]);
-
-
-  const [currentColor, setCurrentColor] = useState(0);
-
-  // store the pair of edges
-  const [connectionPairs, setConnectionPairs] = useState([]);
-
-  const [connectionGroups, setConnectionGroups] = useState([]);
   const groupMapRef = useRef(new Map());
 
-  const[welcomeMessage, setWelcomeMessage] = useState(false);
+  const { clickAudio, errorAudio, connectAudio } = useAudio();
+  const { offset, setOffset, soundBool, setSoundBool, blackDotEffect, setBlackDotEffect } = useSettings();
+  const [showSettings, setShowSettings] = useState(false);
+  const [welcomeMessage, setWelcomeMessage] = useState(false);
+
   useEffect(() => {
-    if(topRowCount === 1 && bottomRowCount === 1) {
+    if (topRowCount === 1 && bottomRowCount === 1) {
       setWelcomeMessage(true);
     }
   }, [topRowCount, bottomRowCount]);
 
   useEffect(() => {
-
     drawConnections(svgRef, connections, connectionPairs, offset);
   }, [connectionGroups, connections, topRowCount, bottomRowCount, connectionPairs, offset]);
 
   useEffect(() => {
-    checkAndAddNewNodes();
+    checkAndAddNewNodes(topRowCount, bottomRowCount, connections, setTopRowCount, setBottomRowCount);
   }, [connections, topRowCount, bottomRowCount]);
 
   useEffect(() => {
-    //console.log("CONNECTION PAIRS:", connectionPairs);
-  }, [connectionPairs]);
-
-  useEffect(() => {
-    calculateProgress();
+    setProgress(calculateProgress(connections, topRowCount, bottomRowCount));
   }, [connections, topRowCount, bottomRowCount]);
 
   useEffect(() => {
     const handleResize = () => {
-      drawConnections(svgRef, connections, connectionPairs, offset); // Pass parameters
+      drawConnections(svgRef, connections, connectionPairs, offset);
     };
-
     window.addEventListener("resize", handleResize);
-
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, [svgRef, connections, connectionPairs]); // Add svgRef to dependencies
+    return () => window.removeEventListener("resize", handleResize);
+  }, [svgRef, connections, connectionPairs]);
 
   useEffect(() => {
     const latestPair = connectionPairs[connectionPairs.length - 1];
@@ -118,45 +78,23 @@ function App() {
     }
   }, [connectionPairs]);
 
-  const checkAndAddNewNodes = () => {
-    const allTopNodesConnected = Array.from({ length: topRowCount }, (_, i) =>
-      connections.some((conn) => conn.nodes.includes(`top-${i}`))
-    ).every(Boolean);
-
-    const allBottomNodesConnected = Array.from(
-      { length: bottomRowCount },
-      (_, i) => connections.some((conn) => conn.nodes.includes(`bottom-${i}`))
-    ).every(Boolean);
-
-    if (allTopNodesConnected || allBottomNodesConnected) {
-      if (allTopNodesConnected) {
-        setTopRowCount((prev) => prev + 1);
-      } else {
-        setBottomRowCount((prev) => prev + 1);
-      }
-    }
-  };
-
-  const createTopRow = (count) => {
-    return Array.from({ length: count }, (_, i) => (
-      <>
-        <TaikoNode
-          key={`top-${i}`}
-          id={`top-${i}`}
-          onClick={() => handleNodeClick(`top-${i}`)}
-          isSelected={selectedNodes.includes(`top-${i}`)}
-          index={i}
-          totalCount={topRowCount}
-          isFaded={count > 1 && i === count - 1}
-          position="top"
-          blackDotEffect={blackDotEffect}
-        />
-      </>
+  const createTopRow = (count) =>
+    Array.from({ length: count }, (_, i) => (
+      <TaikoNode
+        key={`top-${i}`}
+        id={`top-${i}`}
+        onClick={() => handleNodeClick(`top-${i}`)}
+        isSelected={selectedNodes.includes(`top-${i}`)}
+        index={i}
+        totalCount={topRowCount}
+        isFaded={count > 1 && i === count - 1}
+        position="top"
+        blackDotEffect={blackDotEffect}
+      />
     ));
-  };
 
-  const createBottomRow = (count) => {
-    return Array.from({ length: count }, (_, i) => (
+  const createBottomRow = (count) =>
+    Array.from({ length: count }, (_, i) => (
       <TaikoNode
         key={`bottom-${i}`}
         id={`bottom-${i}`}
@@ -166,41 +104,52 @@ function App() {
         totalCount={bottomRowCount}
         isFaded={count > 1 && i === count - 1}
         position="bottom"
-
       />
     ));
-  };
 
   const handleNodeClick = (nodeId) => {
     setErrorMessage("");
-    if(soundBool) {
-      connectAudio.play();
-    }
+    if (soundBool) connectAudio.play();
     if (selectedNodes.includes(nodeId)) {
       setSelectedNodes(selectedNodes.filter((id) => id !== nodeId));
-    } else {
-      if (selectedNodes.length < 2) {
-        const newSelectedNodes = [...selectedNodes, nodeId];
-        setSelectedNodes(newSelectedNodes);
-        if (newSelectedNodes.length === 2) {
-          tryConnect(newSelectedNodes);
-        }
-      }
+    } else if (selectedNodes.length < 2) {
+      const newSelectedNodes = [...selectedNodes, nodeId];
+      setSelectedNodes(newSelectedNodes);
+      if (newSelectedNodes.length === 2) tryConnect(newSelectedNodes);
     }
   };
 
+  const handleToolMenuClick = () => setShowSettings((prev) => !prev);
 
+  const handleClear = () => {
+    setConnectionPairs([]);
+    setConnections([]);
+    setSelectedNodes([]);
+    setBottomRowCount(1);
+    setTopRowCount(1);
+    setEdgeState(null);
+    setErrorMessage("");
+    setProgress(0);
+    setConnectionGroups([]);
+    setCurrentColor(0);
+    groupMapRef.current.clear();
+    console.log(connectionPairs);
+  };
 
-  const handleToolMenuClick = () => {
-    setShowSettings((prev) => !prev);
+  const handleSoundClick = () => {
+    // Toggle the soundBool
+    setSoundBool((prev) => !prev);
+
   };
 
   const handleOffsetChange = (newOffset) => {
     setOffset(newOffset);
     localStorage.setItem("offset", newOffset); //store to localStorage
   };
-  
 
+  const toggleBlackDotEffect = () => {
+    setBlackDotEffect((prev) => !prev);
+  };
 
   const tryConnect = (nodes) => {
     if (nodes.length !== 2) return;
@@ -215,7 +164,7 @@ function App() {
       if(soundBool) {
         errorAudio.play();
       }
-      setErrorMessage("Can't connect two nodes from the same row.");
+      setErrorMessage("Can't connect two vertices from the same row.");
       setSelectedNodes([]);
       return;
     }
@@ -230,7 +179,7 @@ function App() {
       if(soundBool) {
         errorAudio.play();
       }
-      setErrorMessage("These nodes are already connected.");
+      setErrorMessage("These vertices are already connected.");
       setSelectedNodes([]);
       return;
     }
@@ -292,212 +241,54 @@ function App() {
     setSelectedNodes([]);
   };
 
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    setShowNodes(true);
-    setConnections([]);
-    setSelectedNodes([]);
-    setErrorMessage("");
-  };
-
-  const handleClear = () => {
-    setConnectionPairs([]);
-    setConnections([]);
-    setSelectedNodes([]);
-    setBottomRowCount(1);
-    setTopRowCount(1);
-    setEdgeState(null);
-    setErrorMessage("");
-    setProgress(0);
-    setConnectionGroups([]);
-    setCurrentColor(0);
-    groupMapRef.current.clear();
-    console.log(connectionPairs);
-  };
-
-  const handleSoundClick = () => {
-    // Toggle the soundBool
-
-    setSoundBool((prev) => !prev);
-
-  };
-
-  const toggleBlackDotEffect = () => {
-    setBlackDotEffect((prev) => !prev);
-  };
-
- 
-
-  const calculateProgress = () => {
-    let totalPossibleConnections = (topRowCount - 1) * (bottomRowCount - 1);
-    if (totalPossibleConnections % 2 !== 0) {
-      totalPossibleConnections -= 1;
-    }
-    const verticalEdges = connections.length;
-
-    let progressPercentage = totalPossibleConnections > 2 ? (verticalEdges / totalPossibleConnections) * 100 : 0;
-    progressPercentage = Math.min(progressPercentage, 100);
-    setProgress(progressPercentage);
-  };
-
-  const showTooltip = (e) => {
-    setTooltipVisible(true);
-    setTooltipPosition({ x: e.clientX, y: e.clientY });
-  };
-
-  const hideTooltip = () => {
-    setTooltipVisible(false);
-  };
-
-  useEffect(() => {
-    console.log("Updated Connection Groups:", connectionGroups);
-  }, [connectionGroups]);
-
   return (
-    <div
-      style={{
-        textAlign: "center",
-        position: "relative",
-        fontFamily: "Arial, sans-serif",
-      }}
-      className="AppContainer"
-    >
-    <h1 className="title">
-      <a href="https://mineyev.web.illinois.edu/ColorTaiko!/" target="_blank" style={{ textDecoration: "none" }}>
-        <span style={{ color: '#e6194b', backgroundColor: '#000000', fontSize: 'inherit', display: 'inline-block' }}>C</span>
-        <span style={{ color: '#3cb44b', backgroundColor: '#000000', fontSize: 'inherit', display: 'inline-block' }}>o</span>
-        <span style={{ color: '#ffe119', backgroundColor: '#000000', fontSize: 'inherit', display: 'inline-block' }}>l</span>
-        <span style={{ color: '#f58231', backgroundColor: '#000000', fontSize: 'inherit', display: 'inline-block' }}>o</span>
-        <span style={{ color: '#dcbeff', backgroundColor: '#000000', fontSize: 'inherit', display: 'inline-block' }}>r</span>
-        <span style={{ color: '#9a6324', backgroundColor: '#000000', fontSize: 'inherit', display: 'inline-block' }}>T</span>
-        <span style={{ color: '#fabebe', backgroundColor: '#000000', fontSize: 'inherit', display: 'inline-block' }}>a</span>
-        <span style={{ color: '#7f00ff', backgroundColor: '#000000', fontSize: 'inherit', display: 'inline-block' }}>i</span>
-        <span style={{ color: '#f032e6', backgroundColor: '#000000', fontSize: 'inherit', display: 'inline-block' }}>k</span>
-        <span style={{ color: '#42d4f4', backgroundColor: '#000000', fontSize: 'inherit', display: 'inline-block' }}>o</span>
-        <span style={{ color: '#bfef45', backgroundColor: '#000000', fontSize: 'inherit', display: 'inline-block' }}>!</span>
-      </a>
-    </h1>
-
-      
-      {/* Container for progress bar and text */}
-<div style={{ marginTop: '-55px' }}> {/* Adjust this value as needed */}
-  {/* Text above progress bar */}
-  <p style={{
-    color: 'white',
-    fontSize: '14px',
-    textAlign: 'left',
-    marginBottom: '-5px',
-    fontFamily: 'inherit',
-  }}>
-    Can you get to 100%?
-  </p>
-
-  <div
-    className="progress-bar-container"
-    onMouseEnter={showTooltip}
-    onMouseMove={showTooltip}
-    onMouseLeave={hideTooltip}
-  >
-    <div className="progress-bar-fill" style={{ width: `${progress}%` }}>
-      <span className="progress-bar-text">{Math.round(progress)}%</span>
-    </div>
-  </div>
-
-  {/* Formula for progress bar */}
-  <p style={{
-    color: 'white',
-    fontSize: '14px',
-    textAlign: 'left',
-    marginTop: '-7px',
-    marginBottom: '-20px',
-    fontFamily: 'inherit',
-  }}>
-    Progress = <span style={{ display: 'inline-block', verticalAlign: 'middle' }}>
-      <span style={{ display: 'block', textAlign: 'center' }}>verticalEdges</span>
-      <span style={{ display: 'block', borderTop: '1px solid white', paddingTop: '2px', textAlign: 'center' }}>
-        (topRowCount - 1) × (bottomRowCount - 1) - (1 if odd, else 0)
-      </span>
-    </span>
-    <span style={{ marginLeft: '5px' }}>× 100%</span>
-  </p>
-</div>
-    <div className="welcome-message"> 
-    {welcomeMessage && (
-        <div className="fade-message">
-          Connect the nodes!
-        </div>
-    )}
-    </div>
-
-      {tooltipVisible && (
-        <div
-          className="tooltip"
-          style={{ top: tooltipPosition.y + 10, left: tooltipPosition.x + 10 }}
-        >
-          <p>Vertical Edges: {connections.length}</p>
-          <p>Top Nodes: {topRowCount - 1}</p>
-          <p>Bottom Nodes: {bottomRowCount - 1}</p>
-        </div>
+    <div className="app-container">
+      <Title />
+  
+      <ProgressBar
+        progress={progress}
+        connections={connections}
+        topRowCount={topRowCount}
+        bottomRowCount={bottomRowCount}
+      />
+  
+      {welcomeMessage && (
+        <div className="welcome-message fade-message">Connect the nodes!</div>
       )}
-      
-      <div>
+  
       <img
         src={SettingIconImage}
         alt="Settings Icon"
         className="icon"
         onClick={handleToolMenuClick}
       />
-      </div>
+  
       {showSettings && (
-        <SettingsMenu offset={offset} onOffsetChange={handleOffsetChange}
-         soundbool={soundBool} onSoundControl={handleSoundClick}
-         blackDotEffect={blackDotEffect}
-         onToggleBlackDotEffect={toggleBlackDotEffect}/>
+        <SettingsMenu
+          offset={offset}
+          onOffsetChange={handleOffsetChange}
+          soundbool={soundBool}
+          onSoundControl={handleSoundClick}
+          blackDotEffect={blackDotEffect}
+          onToggleBlackDotEffect={toggleBlackDotEffect}
+        />
       )}
-
-      <button
-        onClick={handleClear}
-        style={{
-          position: "absolute",
-          top: "10px",
-          right: "10px",
-          padding: "10px 20px",
-          fontSize: "16px",
-          backgroundColor: "#f44336",
-          color: "white",
-          border: "none",
-          borderRadius: "5px",
-          cursor: "pointer",
-          fontFamily: "inherit", // This will use the font from the parent element
-        }}
-      >
+  
+      <button onClick={handleClear} className="clear-button">
         Clear
       </button>
-      
+  
       <ErrorModal
         className="error-container"
         message={errorMessage}
         onClose={() => setErrorMessage("")}
       />
-
+  
       {showNodes && (
-        <div className="GameBox" style={{ position: "relative" }}>
-          <div className="GameRow">{createTopRow(topRowCount)}</div>
-          <svg
-            ref={svgRef}
-            style={{
-              position: "absolute",
-              top: 0,
-              left: 0,
-              width: "100%",
-              height: "100%",
-              pointerEvents: "none",
-              zIndex: 10,
-            }}
-          />
-          <div className="GameRow" style={{ marginTop: "100px" }}>
-            {createBottomRow(bottomRowCount)}
-          </div>
+        <div className="game-box">
+          <div className="game-row">{createTopRow(topRowCount)}</div>
+          <svg ref={svgRef} className="svg-overlay" />
+          <div className="game-row bottom-row">{createBottomRow(bottomRowCount)}</div>
         </div>
       )}
     </div>
