--- client/src/TaikoNode.css
@@ -1,7 +1,0 @@
-TaikoNode {
-    border: 2px solid red;
-}
-
-.Taiko {
-    border: 2px solid red;
-}
