+++ client/src/components/TaikoNodes/taikoNode.css
@@ -1,0 +1,25 @@
+
+.taiko-node {
+  width: calc(100% / var(--node-count) - 20px);
+  height: calc(100% / var(--node-count) - 20px);
+  max-width: 30px;  
+  max-height: 30px;
+  border-radius: 50%;
+  background-color: white;
+  margin: 5px;
+  cursor: pointer;
+  border: 2px solid black;
+  transition: all 0.5s ease;
+  z-index: 0;
+}
+
+.taiko-node-enter {
+  opacity: 0;
+  transform: scale(0.5);
+  transition: opacity 0.5s ease, transform 0.5s ease;
+}
+
+.taiko-node-enter-active {
+  opacity: 1;
+  transform: scale(1);
+}
