+++ client/src/components/ToolMenu/settingMenu.css
@@ -1,0 +1,35 @@
+.settings-menu {
+    position: fixed;
+    top: 70px;
+    right: 20px;
+    width: 200px;
+    padding: 15px;
+    background: white;
+    border: 1px solid #ccc;
+    border-radius: 8px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+    z-index: 1000;
+  }
+  
+  .settings-menu h3 {
+    color: black;
+    margin: 0 0 10px;
+    font-size: 16px;
+  }
+  
+  .settings-menu label {
+    color: black;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+  }
+  
+  .settings-menu input[type="range"] {
+    width: 100%;
+    margin-top: 8px;
+  }
+  
+  .settings-menu input[type="number"] {
+    margin-top: 5px;
+  }
+
