--- client/src/index.css
@@ -1,223 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Caveat+Brush&family=Caveat:wght@400..700&display=swap');
-
-* {
-  padding: 0px;
-  border: 0px;
-  margin: 0px;
-}
-
-body {
-  background-color: rgb(0, 0, 0);
-  display: flex;
-  justify-content: center;
-  font-family: "Caveat Brush";
-  height: 100vh;
-  overflow: auto;
-  color: white;
-}
-
-.title {
-  font-size: 4rem;
-  margin-top: 10px;
-}
-
-
-.GameBox {
-  height: 95vh;
-  width: 90vw;
-  gap: 350px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  overflow: visible;
-
-
-}
-
-.GameRow {
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  overflow: visible;
-}
-
-
-.GameQuestion {
-  color: aliceblue;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 50px;
-  font-size: 1.8rem;
-  margin-top: 10px;
-  margin-bottom: 20px;
-}
-
-.GameQuestion div {
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 10px;
-}
-
-.GameQuestion input {
-  width: 100px;
-  height: 30px;
-  align-self: center;
-  align-items: center;
-  border-radius: 10px;
-  font-size: 1.2rem;
-  padding-left: 9px;
-}
-
-.taiko-node {
-  width: calc(100% / var(--node-count) - 20px);
-  height: calc(100% / var(--node-count) - 20px);
-  max-width: 30px;  
-  max-height: 30px;
-  border-radius: 50%;
-  background-color: white;
-  margin: 5px;
-  cursor: pointer;
-  border: 2px solid black;
-  transition: all 0.5s ease;
-  z-index: 0;
-}
-
-.taiko-node-enter {
-  opacity: 0;
-  transform: scale(0.5);
-  transition: opacity 0.5s ease, transform 0.5s ease;
-}
-
-.taiko-node-enter-active {
-  opacity: 1;
-  transform: scale(1);
-}
-
-.error-container {
-  color: black;
-}
-
-.progress-bar-container {
-  width: 20%;
-  background-color: #837b7b;
-  border-radius: 5px;
-  margin: 20px 0;
-  overflow: hidden;
-  height: 20px;
-  position: relative;
-}
-
-.progress-bar-fill {
-  background-color: #4caf50;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  transition: width 0.3s ease-in-out;
-  padding-left: 5px; /* Small padding to keep text readable at low progress */
-  color: white;
-  font-weight: bold;
-  border-radius: inherit;
-}
-
-.progress-bar-text {
-  position: absolute;
-  left: 50%;
-  transform: translateX(-50%);
-  font-size: 14px;
-  color: white;
-  white-space: nowrap;
-}
-
-.tooltip {
-  position: absolute;
-  background-color: rgba(0, 0, 0, 0.8);
-  color: white;
-  padding: 8px;
-  border-radius: 5px;
-  font-size: 12px;
-  pointer-events: none;
-  white-space: nowrap;
-  box-shadow: 0px 0px 8px rgba(0, 0, 0, 0.2);
-  z-index: 1000;
-}
-
-.icon {
-  position: absolute;
-  top: 16px;
-  right: 150px;
-  font-size: 24px;
-  cursor: pointer;
-  transition: transform 0.2s ease;
-}
-
-.icon:hover {
-  color: #007bff;
-  transform: scale(1.1); 
-}
-
-.settings-menu {
-  position: fixed;
-  top: 70px;
-  right: 20px;
-  width: 200px;
-  padding: 15px;
-  background: white;
-  border: 1px solid #ccc;
-  border-radius: 8px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
-  z-index: 1000;
-}
-
-.settings-menu h3 {
-  color: black;
-  margin: 0 0 10px;
-  font-size: 16px;
-}
-
-.settings-menu label {
-  color: black;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.settings-menu input[type="range"] {
-  width: 100%;
-  margin-top: 8px;
-}
-
-.settings-menu input[type="number"] {
-  margin-top: 5px;
-}
-
-.fade-message {
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  font-size: 36px;
-  font-family: 'Cinzel Decorative', serif;
-  font-weight: 400;
-  color: #dcdcdc;
-  text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.5);
-  opacity: 0;
-  animation: fadeInOut 3s ease-in-out forwards;
-}
-
-@keyframes fadeInOut {
-  0% {
-    opacity: 0;
-  }
-  25% {
-    opacity: 1;
-  }
-  75% {
-    opacity: 1;
-  }
-  100% {
-    opacity: 0;
-  }
-}
