+++ client/src/main.jsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App.jsx'
-import './index.css'
+import './styles/global.css'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
