+++ client/src/styles/global.css
@@ -1,0 +1,135 @@
+@import url('https://fonts.googleapis.com/css2?family=Caveat+Brush&family=Caveat:wght@400..700&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=STIX+Two+Math&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Cinzel+Decorative:wght@700&display=swap');
+
+* {
+  padding: 0px;
+  border: 0px;
+  margin: 0px;
+}
+
+body {
+  background-color: rgb(0, 0, 0);
+  display: flex;
+  justify-content: center;
+  font-family: "Caveat Brush";
+  height: 100vh;
+  overflow: auto;
+  color: white;
+}
+
+.title {
+  font-size: 4rem;
+  margin-top: 10px;
+}
+
+.error-container {
+  color: black;
+}
+
+.tooltip {
+  position: absolute;
+  background-color: rgba(0, 0, 0, 0.8);
+  color: white;
+  padding: 8px;
+  border-radius: 5px;
+  font-size: 12px;
+  pointer-events: none;
+  white-space: nowrap;
+  box-shadow: 0px 0px 8px rgba(0, 0, 0, 0.2);
+  z-index: 1000;
+}
+
+.icon {
+  position: absolute;
+  top: 16px;
+  right: 150px;
+  font-size: 24px;
+  cursor: pointer;
+  transition: transform 0.2s ease;
+}
+
+.icon:hover {
+  color: #007bff;
+  transform: scale(1.1); 
+}
+
+.fade-message {
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  font-size: 36px;
+  font-family: 'Cinzel Decorative', serif;
+  font-weight: 400;
+  color: #dcdcdc;
+  text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.5);
+  opacity: 0;
+  animation: fadeInOut 3s ease-in-out forwards;
+}
+
+@keyframes fadeInOut {
+  0%, 100% { opacity: 0; }
+  25%, 75% { opacity: 1; }
+}
+
+/* App.css */
+.app-container {
+  text-align: center;
+  position: relative;
+  font-family: Arial, sans-serif;
+}
+
+.welcome-message {
+  /* Additional fade animation if needed */
+}
+
+.icon {
+  cursor: pointer;
+  /* Other styling as needed */
+}
+
+.clear-button {
+  position: absolute;
+  top: 10px;
+  right: 10px;
+  padding: 10px 20px;
+  font-size: 16px;
+  background-color: #f44336;
+  color: white;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+}
+
+.svg-overlay {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  pointer-events: none;
+  z-index: 10;
+}
+
+.game-box {
+  position: relative;
+  height: 95vh;
+  width: 90vw;
+  gap: 350px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  overflow: visible;
+}
+
+.game-row {
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  overflow: visible;
+}
+
+.bottom-row {
+  margin-top: 100px;
+}
