--- package.json
@@ -1,6 +1,0 @@
-{
-  "dependencies": {
-    "@xyflow/react": "^12.3.1",
-    "reactflow": "^11.11.4"
-  }
-}
